requirements.txt
PyYAML==6.0.2
garth==0.5.2
pytest==8.3.4
pytest-cov==6.0.0